feat: Add form_radio_porschecamp.html5 template
This commit adds the form_radio_porschecamp.html5 template, which is used for rendering a radio button field in the Porsche Camps module. The template includes HTML markup and PHP code to display a list of camps fetched from the database and allow the user to select one. It also includes JavaScript code to handle the selection and styling of the radio buttons.

Note: No newline at end of file